Ubah constructor Square ke int.
Karena inputnya int.

+++ square.cpp
@@ -1,9 +1,9 @@
 #include "square.h"
 Square::Square()
 {
-	Square(1.0);
+	Square(1);
 }
-Square::Square(float sisi)
+Square::Square(int sisi)
 {
 	panjang=sisi;
 	lebar=sisi;
